efw-downloader: efw-proto: add constructor, destructor, bind, unbind functions
In Fireworks protocol, response frame from target device arrives at offset
0xecc080000000 on 1394 OHCI controller and the maximum size is 0x200U. It
seems to be inspired by Function Control Protocol in IEC 61883-1.

This commit adds functions to construct and destruct the instance of
EfwProto object, to bind and unbind the Fireworks protocol to the the given
instance of HinawaFwNode. When calling the bind function, the instance
requests Linux firewire subsystem to allocate the range of address.
It's possible to encounter conflict in the case that the range of address
is also allocated by ALSA fireworks driver, thus this tool can not be used
when the driver is loaded.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>