efw-downloader: start a new project to operate on-board flash memory for Fireworks board module
Fireworks board module was designed by Echo Digital Audio corporation. The
board module has on-board flash memory to store firmware blob and session
data. The contents of flash memory can be operated by software by a pair
of asynchronous transactions defined by Echo Digital Audio corporation.

The firmware blob is used by below ICs:

 * Texus Instruments TSB43Cx43A (IceLynx Micro)
 * Texus Instruments TMS320C6713B (DSP)
 * Xilinx Spartan XC35250E (FPGA)

All of Fireworks models use the IceLynx Micro for packet processing on
IEEE 1394 bus. On the other hand, a part of the models use the DSP and
the others use FPGA. The vendors of models ships several versions of
a set of firmware.

This project produces a command-line tool named as 'efw-downloader' to
operate contents of on-board flash memory for firmware of the above ICs.
This software is licensed under the GNU public license v3.0 or later.
Meson build system is used to build the tool.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>