efw-downloader: efw-proto: add function to finish transaction for command frame
In Fireworks protocol, asynchronous transaction is used to deliver command
frame from software to device. The transaction is done to offset
0xecc000000000 with variable length of data up to 0x200.

The transaction includes header and data in its payload. The header
includes sequence field. The target device increments the value of
sequence field, then transfers response frame with the value so that
software can correspond the response frame with command frame according
to the sequence numbers. The header also includes version field but
version 1 is always used as long as I investigate. The header includes
category and command field to determine the type of data. The status
field is used to report operation result from the target device.

This commit adds a function to finish transaction for command frame.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>