efw-downloader: efw-proto: emit responded signal at receiving response
HinawaFwResp has class virtual method for applications to handle received
frame on 1394 OHCI controller by overriding the method with own function
implementation.

This commit overrides the class virtual method and emits 'responded'
signal with the data of received response frame. The content of frame
is aligned to host-endianness, then passed to signal handlers.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>