efw-downloader: file-cntr: add parser for data binary shipped by Echo Digital Audio corporation
Echo Digital Audio corporation designed data format of firmware. The
format seems to be suitable for their SDK for customers and closed.

FFADO project implements the format parser in libffado2. Although the
implementation is not enough in several points, it's good point to guess
the format.

I captured asynchronous packet by bus analyzer during updating firmware by
Windows driver, aggregated the content of packet by script, then compared
it to content of firmware. Some points are cleared (but not all).

This commit adds parser for data binary shipped by Echo Digital Audio
corporation. I note that driver package shipped by vendors for Fireworks
based devices includes no EULA. I think this kind of reimplementation
is not limited in the case.

The content of file consists of two parts; header after magic bytes, and
payload. The header includes metadata about the data of payload. There are
four kind of data; for DSP, for IceLynx Micro, for arbitrary data, and
for FPGA.

The content except for magic bytes is hexadecimal string,
therefore it's possible to print out by simple dump tool; cat(1) and
less(1). The header includes CRC32 value of the decoded data and checksum
of every bytes of the decoded data.

When the header includes a flag, the CRC32 value is written to the end of
area for firmware.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>