efw-downloader: efw-proto: add function to finish a pair of transactions
In Fireworks protocol, the target device increments sequence number in
command frame transferred by software, then transfers response frame
with the number so that software can correspond the response frame to
the command frame.

This commit adds function to finish the pair of transactions for command
and response frames. The frames are matched according to sequence number.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>